v5 run settings

+++ DGMR-SO/utils/losses.py
@@ -11,7 +11,7 @@ class Loss_hing_disc():
         l2 = tf.nn.relu(1. + score_generated)
         loss += tf.reduce_mean(l2)  # , axis=list(range(1, len(l2.shape)))
         tf.print("Debugging: Disc Loss: ", loss)
-        return tf.cast(loss, tf.float32)
+        return loss
 
 class Loss_hing_gen():
     def __init__(self) -> None:
@@ -23,4 +23,4 @@ class Loss_hing_gen():
             tf.reduce_mean(
                 score_generated)  # ,axis=list(range(1, len(score_generated.shape)))
         tf.print("Debugging, Gen Loss: ", loss)
-        return tf.cast(loss, tf.float32)
+        return loss
